Add autopwn details to adobe_cooltype_sing
This is my favorite browser module because it's drop-dead easy to test
with across a ton of different browsers. For more of my gushing on this,
see:

https://community.rapid7.com/community/metasploit/blog/2011/10/10/client-side-exploit-testing-with-adobecooltypesing

This should be landed to make testing browser_autopwn pretty easy in any
given situation, since it's easy to ensure that at least your Adobe
Reader is unpatched, and doesn't depend on particular (sometimes hard to
find) java versions or browser versions.